requirements.txt
# Core dependencies
firecrawl-py==1.5.0
python-dotenv==1.1.1
requests==2.32.4
pydantic==2.11.7

# Data processing
PyYAML==6.0.2

# Development and testing (optional)
pytest==8.3.3
black==24.10.0
flake8==7.1.1

# Documentation
Sphinx==8.1.3 